requirements.txt
aiohappyeyeballs==2.4.8
aiohttp==3.11.13
aiohttp-retry==2.9.1
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.1.0
bcrypt==4.3.0
boto3==1.37.7
botocore==1.37.7
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cssselect2==0.7.0
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
fonttools==4.56.0
frozenlist==1.5.0
gunicorn==23.0.0
idna==3.10
Jinja2==3.1.5
jmespath==1.0.1
MarkupSafe==3.0.2
multidict==6.1.0
packaging==24.2
pdfkit==1.0.0
pillow==11.1.0
propcache==0.3.0
psycopg==3.2.5
psycopg-binary==3.2.5
pycparser==2.22
pydyf==0.11.0
PyJWT==2.9.0
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-http-client==3.3.7
pytz==2025.1
requests==2.32.3
s3transfer==0.11.4
sendgrid==6.11.0
six==1.17.0
sqlparse==0.5.3
starkbank-ecdsa==2.2.0
tinycss2==1.4.0
tinyhtml5==2.0.0
twilio==9.4.6
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
weasyprint==64.1
webencodings==0.5.1
yarl==1.18.3
zopfli==0.2.3.post1
